refactor(core): move injectors away from render item component
And abstract render component

re #28
