+++ libs/core/src/lib/injectors/local-injector-map.ts
@@ -3,8 +3,10 @@ import {
   InjectionToken,
   Injector,
   StaticProvider,
+  Type,
 } from '@angular/core';
 
+import { RenderComponent } from '../render-component';
 import { InjectorMap } from '../types';
 
 export type LocalGetInjectorToken = () => Injector;
@@ -12,7 +14,6 @@ export type LocalGetInjectFlagsToken = InjectFlags;
 export type LocalGetComponentToken = () => any;
 export type LocalGetConfigToken = () => any;
 export type LocalIsConfigValidToken = () => boolean;
-export type LocalGetRenderItemToken = () => any;
 
 export const LOCAL_GET_INJECTOR = new InjectionToken<LocalGetInjectorToken>(
   'LOCAL_GET_INJECTOR',
@@ -34,17 +35,13 @@ export const LOCAL_GET_CONFIGURATION_VALID = new InjectionToken<
   LocalIsConfigValidToken
 >('LOCAL_GET_CONFIGURATION_VALID');
 
-export const LOCAL_GET_RENDER_ITEM = new InjectionToken<
-  LocalGetRenderItemToken
->('LOCAL_GET_RENDER_ITEM');
-
 export const LOCAL_INJECTOR_MAP: InjectorMap = {
   getinjector: LOCAL_GET_INJECTOR,
   injectflags: LOCAL_GET_INJECT_FLAGS,
   getcomponent: LOCAL_GET_COMPONENT,
   getconfiguration: LOCAL_GET_CONFIGURATION,
   isconfigurationvalid: LOCAL_GET_CONFIGURATION_VALID,
-  getrenderitem: LOCAL_GET_RENDER_ITEM,
+  getrendercomponent: RenderComponent as Type<RenderComponent>,
 };
 
 export function getLocalProviders(data: {
@@ -52,7 +49,6 @@ export function getLocalProviders(data: {
   getComponent: LocalGetComponentToken;
   getConfig: LocalGetConfigToken;
   isConfigValid: LocalIsConfigValidToken;
-  getRenderItem: LocalGetRenderItemToken;
 }): StaticProvider[] {
   return [
     {
@@ -74,10 +70,6 @@ export function getLocalProviders(data: {
     {
       provide: LOCAL_GET_CONFIGURATION_VALID,
       useValue: data.getConfig,
-    },
-    {
-      provide: LOCAL_GET_RENDER_ITEM,
-      useValue: data.getRenderItem,
     },
   ];
 }
